Optical Propagation Methods

Optical Propagation Methods

Here are some useful optical propagation methods.

OpticalPropagation.angularspectrum — Function.

angularspectrum(Uin, d, λ, lx, ly)

calculate the propagation light field based on the angular spectrum.

Parameters

Uin::AbstractArray{<:Number,2}: Complex array of input complex amplitude.

d::Real: Distance to propagate in metres.

λ::Real: Wavelength of light to propagate.

lx::Real: The physical size of the input data along the x-axis.

ly::Real: The physical size of the input data along the y-axis.

Returns

::Array{<:Number,2}: Complex amplitude data after propagation.

source

angularspectrum(Uin, d)

calculate the propagation light field based on the angular spectrum.

Parameters

Uin::MonoLightField2D: Light field data in input plane.

d::Unitful.Length: Distance to propagate.

Returns

::MonoLightField2D: Light field data after propagation.

source

OpticalPropagation.fresnel1 — Function.

fresnel1(Uin, d, λ, lx, ly)

calculate the propagation light field based on the Fresnel diffraction with single Fourier transform.

Parameters

Uin::AbstractArray{<:Number,2}: Complex array of input complex amplitude.

d::Real: Distance to propagate in metres.

λ::Real: Wavelength of light to propagate.

lx::Real: The physical size of the input data along the x-axis.

ly::Real: The physical size of the input data along the y-axis.

Returns:

(Uout, (lxo, lyo))

Uout::Array{<:Number,2}: Complex amplitude data after propagation.

lxo::Real: The physical size of the diffraction data along the x-axis.

lyo::Real: The physical size of the diffraction data along the y-axis.

source

fresnel1(Uin, d)

calculate the propagation light field based on the Fresnel diffraction with single Fourier transform.

Parameters

Uin::MonoLightField2D: Light field data in input plane.

d::Real: Distance to propagate.

Returns

::MonoLightField2D: Light field data after propagation.

source

OpticalPropagation.fresnel2 — Function.

fresnel2(Uin, d, λ, lx, ly)

calculate the propagation light field based on the Fresnel diffraction with double Fourier transform.

Parameters

Uin::AbstractArray{<:Number,2}: Complex array of input complex amplitude.

d::Real: Distance to propagate in metres.

λ::Real: Wavelength of light to propagate.

lx::Real: The physical size of the input data along the x-axis.

ly::Real: The physical size of the input data along the y-axis.

Returns

::Array{<:Number,2}: Complex amplitude data after propagation.

source

fresnel2(Uin, d)

calculate the propagation light field based on the Fresnel diffraction with double Fourier transform.

Parameters

Uin::MonoLightField2D: Light field data in input plane.

d::Real: Distance to propagate.

Returns

::MonoLightField2D: Light field data after propagation.

source