Home

OpticalPropagation.jl

OpticalPropagation.angular_spectrum

OpticalPropagation.fresnel_diffraction_double

OpticalPropagation.fresnel_diffraction_single

OpticalPropagation.angular_spectrum — Method.

angular_spectrum(data, d, λ, lx, ly)

calculate the propagation light field based on the angular spectrum.

source

OpticalPropagation.fresnel_diffraction_double — Method.

fresnel_diffraction_double(data, d, λ, lx, ly)

calculate the propagation light field based on the Fresnel diffraction with double Fourier transform.

source

OpticalPropagation.fresnel_diffraction_single — Method.

fresnel_diffraction_single(data, d, λ, lx, ly)

calculate the propagation light field based on the Fresnel diffraction with single Fourier transform.

source