Home

LightPropagation.jl

LightPropagation.angular_spectrum

LightPropagation.fresnel_diffraction_double

LightPropagation.fresnel_diffraction_single

LightPropagation.angular_spectrum — Method.

angular_spectrum(data, d, λ, lx, ly)

calculate the propagation light field based on the angular spectrum.

source

LightPropagation.fresnel_diffraction_double — Method.

fresnel_diffraction_double(data, d, λ, lx, ly)

calculate the propagation light field based on the Fresnel diffraction with double Fourier transform.

source

LightPropagation.fresnel_diffraction_single — Method.

fresnel_diffraction_single(data, d, λ, lx, ly)

calculate the propagation light field based on the Fresnel diffraction with single Fourier transform.

source